microservices/ai_ameasure/requirements.txt
# FastAPI関連
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# データ処理
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
PyYAML==6.0.1

# 地理空間データ処理
pyproj==3.6.1
shapely==2.0.2
geopandas==0.14.1
folium==0.15.1
srtm.py==0.3.7
rasterio==1.3.9
gpxpy==1.6.1

# 可視化
matplotlib==3.7.2
seaborn==0.12.2
japanize-matplotlib==1.1.3

# その他
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# テスト
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2